Split UDP protocol tests by contract
